refactor mock plugin code
