+++ shenyu-plugin/shenyu-plugin-mock/src/main/java/org/apache/shenyu/plugin/mock/generator/RandomIntGenerator.java
@@ -17,10 +17,10 @@
 
 package org.apache.shenyu.plugin.mock.generator;
 
-import java.util.List;
-import org.apache.shenyu.plugin.base.mock.Generator;
 import org.apache.shenyu.plugin.mock.util.RandomUtil;
 import org.apache.shenyu.spi.Join;
+
+import java.util.List;
 
 /**
  * Random int value generator in the specified range.
