+++ shenyu-plugin/shenyu-plugin-mock/src/main/resources/META-INF/shenyu/org.apache.shenyu.plugin.mock.generator.Generator
@@ -13,6 +13,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 bool=org.apache.shenyu.plugin.mock.generator.BoolGenerator
 current=org.apache.shenyu.plugin.mock.generator.CurrentTimeGenerator
 email=org.apache.shenyu.plugin.mock.generator.EmailGenerator
